Amelioration du back

+++ back/src/main/java/com/openclassrooms/starterjwt/services/UserService.java
@@ -1,8 +1,11 @@
 package com.openclassrooms.starterjwt.services;
 
+import com.openclassrooms.starterjwt.exception.NotFoundException;
 import com.openclassrooms.starterjwt.models.User;
 import com.openclassrooms.starterjwt.repository.UserRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 public class UserService {
@@ -12,11 +15,27 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public void delete(Long id) {
+    public User findById(Long id) {
+        return this.userRepository.findById(id).orElseThrow(NotFoundException::new);
+    }
+
+    public void delete(Long id, String authenticatedEmail) {
+        User user = this.findById(id);
+        if (!user.getEmail().equals(authenticatedEmail)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         this.userRepository.deleteById(id);
     }
 
-    public User findById(Long id) {
-        return this.userRepository.findById(id).orElse(null);
+    public User findByEmail(String email) {
+        return this.userRepository.findByEmail(email).orElse(null);
+    }
+
+    public boolean existsByEmail(String email) {
+        return this.userRepository.existsByEmail(email);
+    }
+
+    public User save(User user) {
+        return this.userRepository.save(user);
     }
 }
